o always replay journals when their publish region is mounted o only do a full check when mnt count or time are exceeded or -f o add some messages to let the user know what is going on o don't give replay access to o2fsck_state, it doesn't use it o clean up the setting and clearing of bits in the global bitmap o update the lastcheck field in the superblock and write it out
Original-revision: 462

+++ fsck.ocfs2/include/problem.h
@@ -33,7 +33,7 @@
 /* returns non-zero for yes and zero for no.  The caller is expected to
  * provide a thorough description of the state and the action that will
  * be taken depending on the answer.  Without \n termination. */
-int prompt(o2fsck_state *ost, unsigned flags, unsigned code, 
+int prompt(o2fsck_state *ost, unsigned flags, uint16_t code, 
 	   const char *fmt, ...)
 	 __attribute__ ((format (printf, 4, 5)));
